+++ src/design/plone/ioprenoto/restapi/services/booking_schema/get.py
@@ -1,7 +1,18 @@
 # -*- coding: utf-8 -*-
+from plone import api
 from redturtle.prenotazioni.restapi.services.booking_schema.get import (
     BookingSchema as BaseService,
 )
+from zope.component import queryMultiAdapter
+
+
+try:
+    from design.plone.iocittadino.interfaces import IDesignPloneIocittadinoLayer
+    from design.plone.iocittadino.interfaces import IUserStore
+
+    WITH_IOCITTADINO = True
+except ImportError:
+    WITH_IOCITTADINO = False
 
 
 class BookingSchema(BaseService):
@@ -31,3 +42,22 @@ class BookingSchema(BaseService):
             "description": "Inserire il nome completo",
         },
     }
+
+    def reply(self):
+        data = super().reply()
+        if WITH_IOCITTADINO and IDesignPloneIocittadinoLayer.providedBy(self.request):
+            # se è configurato iocittadino fare onceonly con i dati di iocittadino
+            if not api.user.is_anonymous():
+                user = api.user.get_current()
+                userstore = queryMultiAdapter(
+                    (self.context, user, self.request), IUserStore
+                )
+                defaults = userstore.get()
+                for field in data["fields"]:
+                    if field["name"] in userstore.user_properties:
+                        if defaults.get(field["name"]):
+                            field["value"] = defaults[field["name"]]
+                        field["readonly"] = (
+                            field["name"] in userstore.strict_user_properties
+                        )
+        return data